Perhaps mysocket.c is compliant to Maekawa's IPv6 library.
That is, mysocket.c contains the following function of IPv4 version:
ox_connect
ox_listen
ox_accept ( = mysocketAccept)
ox_getport

+++ src/ox_toolkit/mysocket.h
@@ -1,12 +1,12 @@
 /* -*- mode: C; coding: euc-japan -*- */
-/* $OpenXM: OpenXM/src/ox_toolkit/mysocket.h,v 1.1 1999/12/09 22:44:56 ohara Exp $ */
+/* $OpenXM: OpenXM/src/ox_toolkit/mysocket.h,v 1.2 2000/10/10 05:23:20 ohara Exp $ */
 
 #ifndef _MYSOCKET_H_
 
 #define _MYSOCKET_H_
 
-int mysocketListen(char *hostname, short *portp);
-int mysocketOpen(char* hostname, short port);
+int mysocketListen(char *hostname, int *portp);
+int mysocketOpen(char* hostname, int port);
 int mysocketAccept(int s_waiting);
 #if 0
 int mypipe(char *program, int fd1, int fd2);
